requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
idna==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.2
openai==1.3.0
